+++ modules/device/scripts/management/default.nix
@@ -1,4 +1,4 @@
-{ config, lib, pkgs, ... }:
+{ config, lib, inputs, pkgs, ... }:
 let
   cfg = config.scripts.management;
 
@@ -12,51 +12,48 @@ let
       nix flake update /etc/nixos $2
     ;;
     "apply")
-      echo "Applying Device Settings..."
-      sudo nixos-rebuild switch --flake /etc/nixos#
-      printf "\n"
-      echo "Applying User Settings..."
-      nix build /etc/nixos#homeConfigurations.$USER.activationPackage && ./result/activate && rm -rf ./result
+      case $2 in
+      "")
+        nixos apply device
+        printf "\n"
+        nixos apply user
+      ;;
+      "device")
+        echo "Applying Device Settings..."
+        case $3 in
+        "") sudo nixos-rebuild switch --flake /etc/nixos#;;
+        "--boot") sudo nixos-rebuild boot --flake /etc/nixos#;;
+        "--test") sudo nixos-rebuild test --flake /etc/nixos#;;
+        "--check") nixos-rebuild dry-activate --flake /etc/nixos#;;
+        *) echo "error: Unknown option $3" && exit 125;;
+        esac
+      ;;
+      "user")
+        echo "Applying User Settings..."
+        nix build /etc/nixos#homeConfigurations.$USER.activationPackage && ./result/activate && rm -rf ./result
+      ;;
+      *) echo "error: Unknown option $2" && exit 125;;
+      esac
     ;;
-    "apply-device")
-      echo "Applying Device Settings..."
-      if [ -z "$2" ]; then
-        sudo nixos-rebuild switch --flake /etc/nixos#
-      elif [ $2 = "--boot" ]; then
-        sudo nixos-rebuild boot --flake /etc/nixos#
-      elif [ $2 = "--test" ]; then
-        sudo nixos-rebuild test --flake /etc/nixos#
-      elif [ $2 = "--check" ]; then
-        nixos-rebuild dry-activate --flake /etc/nixos#
-      else
-        echo "Unknown option $2"
-      fi
-    ;;
-    "apply-user")
-      echo "Applying User Settings..."
-      nix build /etc/nixos#homeConfigurations.$USER.activationPackage && ./result/activate && rm -rf ./result
-   ;;
-   "iso")
-     echo "Building $2 ISO file..."
-     nix build /etc/nixos#installMedia.$2.config.system.build.isoImage && echo "The image is located at ./result/iso/nixos.iso"
-     if [ -z "$3" ]; then
-       echo "The --burn option can be used to burn the image to a USB"
-     elif [ $3 = "--burn" ]; then
-       if [ -z "$4" ]; then
-         echo "Expected a path to USB drive following --burn"
-       else
-         sudo dd if=./result/iso/nixos.iso of=$4 status=progress bs=1M
-       fi
-     else
-       echo "Unexpected option $3"
-     fi
+    "iso")
+      echo "Building $2 ISO file..."
+      nix build /etc/nixos#installMedia.$2.config.system.build.isoImage && echo "The image is located at ./result/iso/nixos.iso"
+      case $3 in
+      "") echo "The --burn option can be used to burn the Image to a USB";;
+      "--burn")
+        case $4 in
+        "") echo "error: Expected a path to USB Drive following --burn" && exit 125;;
+        *) sudo dd if=./result/iso/nixos.iso of=$4 status=progress bs=1M;;
+        esac
+      ;;
+      *) echo "error: Unknown option $3" && exit 125;;
+      esac
     ;;
     "shell")
-      if [ -z "$2" ]; then
-        nix develop /etc/nixos
-      else
-        nix develop /etc/nixos#$2
-      fi
+      case $2 in
+      "") nix develop /etc/nixos;;
+      *) nix develop /etc/nixos#$2;;
+      esac
     ;;
     "list")
       nix-store -q -R /run/current-system | sed -n -e 's/\/nix\/store\/[0-9a-z]\{32\}-//p' | sort | uniq
@@ -78,18 +75,39 @@ let
       echo "Running De-Duplication..."
       nix store optimise
     ;;
+    "secret")
+      case $2 in
+      "list") tree -C --noreport ${inputs.secrets};;
+      "edit")
+        case $3 in
+        "") echo "error: Expected a path to Secret following edit command" && exit 125;;
+        *)
+          echo "Editing Secret $3..."
+          git clone https://github.com/maydayv7/secrets secrets.bak && pushd secrets.bak
+          nano ./$3
+          git add .
+          read -p "Enter comment: " comment
+          git commit -m "$(echo $comment)"
+          git push
+          popd && rm -rf secrets.bak
+          nix flake lock --update-input secrets /etc/nixos
+        ;;
+        esac
+      ;;
+      *) echo "error: Unknown option $2" && exit 125;;
+      esac
+    ;;
     *)
-      echo "########################## Tool for NixOS System Management ##########################"
-      echo "Usage :-"
-      echo "update                        - Updates Nix Flake Inputs"
-      echo "apply                         - Applies both Device and User Configuration"
-      echo "apply-device [ --'option' ]   - Applies Device Configuration"
-      echo "apply-user                    - Applies User Home configuration"
-      echo "iso 'image' [ --burn 'path' ] - Builds Install Media and optionally burns it to USB"
-      echo "shell [ 'name' ]              - Opens desired Nix Developer Shell"
-      echo "list                          - Lists all Installed Packages"
-      echo "save                          - Saves Configuration State to Repository"
-      echo "clean                         - Garbage Collects and Hard-Links Nix Store"
+      echo "## Tool for NixOS System Management ##"
+      echo "# Usage #"
+      echo "  update [ --'option' ]              - Updates Nix Flake Inputs"
+      echo "  apply  [ 'choice' [ --'option' ] ] - Applies Device and User Configuration"
+      echo "  iso 'variant' [ --burn 'path' ]    - Builds Install Media"
+      echo "  shell [ 'name' ]                   - Opens desired Nix Developer Shell"
+      echo "  list                               - Lists all Installed Packages"
+      echo "  save                               - Saves Configuration State to Repository"
+      echo "  clean                              - Garbage Collects and Hard-Links Nix Store"
+      echo "  secret 'choice' [ 'path' ]         - Manages system Secrets"
     ;;
     esac
   '';
@@ -102,8 +120,8 @@ in rec
     default = false;
   };
 
-  config = lib.mkIf (cfg == true)
+  config = lib.mkIf cfg
   {
-    environment.systemPackages = [ script ]; 
+    environment.systemPackages = with pkgs; [ script tree ];
   };
 }