+++ modules/user/shell/git/default.nix
@@ -1,6 +1,7 @@
 { config, lib, inputs, ... }:
 let
   cfg = config.shell.git;
+  key = (builtins.readFile "${inputs.secrets}/gpg/key");
 in rec
 {
   options.shell.git =
@@ -25,16 +26,10 @@ in rec
       type = lib.types.str;
       default = "maydayv7@gmail.com";
     };
-
-    key = lib.mkOption
-    {
-      type = lib.types.str;
-      description = "GPG Signing Key";
-    };
   };
 
   ## User Git Configuration ##
-  config = lib.mkIf (cfg.enable == true)
+  config = lib.mkIf cfg.enable
   {
     programs.git =
     {
@@ -61,34 +56,9 @@ in rec
       userEmail = cfg.userMail;
       signing =
       {
-        key = cfg.key;
+        inherit key;
         signByDefault = true;
       };
     };
-
-    ## Import Keys ##
-    # GPG Keys
-    home.activation.importGPGKeys = inputs.home-manager.lib.hm.dag.entryAfter ["writeBoundary"]
-    ''
-      echo "Importing GPG Keys..."
-      $DRY_RUN_CMD mkdir -p ~/.gnupg $VERBOSE_ARG
-      $DRY_RUN_CMD gpg --import ${inputs.secrets}/gpg/public.gpg $VERBOSE_ARG
-      $DRY_RUN_CMD gpg --import ${inputs.secrets}/gpg/private.gpg $VERBOSE_ARG
-    '';
-
-    # SSH Keys
-    home.activation.importSSHKeys = inputs.home-manager.lib.hm.dag.entryAfter ["writeBoundary"]
-    ''
-      FILE=~/.ssh
-      if [ -e "$FILE" ];
-      then
-        echo "SSH Keys already imported"
-      else
-        echo "Importing SSH Keys..."
-        $DRY_RUN_CMD cp ${inputs.secrets}/ssh ~/.ssh -r $VERBOSE_ARG
-        $DRY_RUN_CMD chmod 400 ~/.ssh/id_ed25519
-        $DRY_RUN_CMD ssh-add ~/.ssh/id_ed25519
-      fi
-    '';
   };
 }